experiments/eval-requirements.txt
# Python deps for local evaluation & visualization only.
# Not named requirements.txt — Render auto-detects Python from that filename in the service root.
# Install: pip install -r experiments/eval-requirements.txt
oracledb>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
matplotlib>=3.7.0
numpy>=1.24.0
